+++ NEWS.md
@@ -29,6 +29,7 @@ New library functions
 ---------------------
 
 * New function `isgreater(a, b)` defines a descending total order where unorderable values and missing are ordered smaller than any regular value.
+* Two argument methods `findmax(f, domain)`, `argmax(f, domain)` and the corresponding `min` versions ([#27613]).
 
 New library features
 --------------------